DSO-688 disable ip masking for app insights
app insights is the main tool used for app services to analyse requests.
We could ship the logs to LA workspace but the app insights are already implemented
we just need to see the IPs which are by default masked.

This is needed to analyse client ip addresses so we can restrict access to all app services
going forward.